'Identifier' classes for generating an unique identifier of the feed entry.
This depends also on the type of feed, e.g. if an feed entry has no 'Id' element (like RSS1 feeds do), it's obviously not possible to use an IdIdentifier. The identifier to use has therefor injected by the frontend to the Core constructor. TitleIdentifier or LinkIdentifier should always work though.